% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{pga_all_time_record_categories}
\alias{pga_all_time_record_categories}
\title{List all-time record categories}
\usage{
pga_all_time_record_categories(tour = "R")
}
\arguments{
\item{tour}{Character. Tour code. Defaults to \code{"R"}.}
}
\value{
A tibble of record IDs with category / subcategory labels.
}
\description{
List all-time record categories
}
