% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{pga_news_franchises}
\alias{pga_news_franchises}
\title{Get news franchise/category list}
\usage{
pga_news_franchises(tour = "R")
}
\arguments{
\item{tour}{Character. Tour code. Defaults to \code{"R"}.}
}
\value{
A tibble with franchise and label columns.
}
\description{
Returns available news categories for filtering with \code{\link[=pga_news]{pga_news()}}.
}
\examples{
\dontrun{
pga_news_franchises()
}
}
