% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_profile.R
\name{pga_player_tournament_status}
\alias{pga_player_tournament_status}
\title{Get player tournament status}
\usage{
pga_player_tournament_status(player_id)
}
\arguments{
\item{player_id}{Character. Player ID.}
}
\value{
A tibble with one row, or empty if the player is not in
the current tournament.
}
\description{
Returns a player's status in the current tournament (if playing),
including position, score, and tee time.
}
\examples{
\dontrun{
pga_player_tournament_status("52955")
}
}
