% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{pga_university_rankings}
\alias{pga_university_rankings}
\title{Get PGA TOUR University rankings}
\usage{
pga_university_rankings(year = NULL, week = NULL)
}
\arguments{
\item{year}{Integer or \code{NULL}. Season year.}

\item{week}{Integer or \code{NULL}. Ranking week.}
}
\value{
A tibble with one row per player.
}
\description{
Get PGA TOUR University rankings
}
