% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgatouR-package.R
\docType{package}
\name{pgatouR-package}
\alias{pgatouR}
\alias{pgatouR-package}
\title{pgatouR: R Client for the PGA Tour API}
\description{
Access PGA Tour leaderboards, player stats, scorecards, shot tracking data, tee times, FedExCup standings, odds, and more. Wraps the PGA Tour GraphQL and REST APIs, returning tidy tibbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WalrusQuant/pgatouR}
  \item Report bugs at \url{https://github.com/WalrusQuant/pgatouR/issues}
}

}
\author{
\strong{Maintainer}: Adam Wickwire \email{adam@adamwickwire.com}

}
\keyword{internal}
